vk: Split the functions into three structs
Split the functions into three structs, one for functions retreived
without an instance, one for functions retrieved with the instance and a
final one for functions retrieved with the device. Most of the code only
needs the last struct so it’s not too much hassle to manage. This will
make it easier to work with in Rust because if it was all one struct we
would need to have partially initialised state after each part is
loaded. With the split approach we can just make a normal constructor
for each struct that either completely succeeds or completely fails. It
also makes it a bit more clear that functions from vk_vr_library can’t
be used when an external device is used.