script: Move the line processing into vr_stream
The vr_stream is now constructed with a pointer to the vr_source. It now
handles opening the source, parsing the lines and doing the token
replacements. This helps the balance with the amount of code in
vr_script will also make it easier to port vr-stream and vr-source to
Rust.