stream: Pass the line buffer as an argument
Instead of having the stream own a buffer and return slice references to
it when reading lines, the line buffer is now passed in as a mutable
reference argument much like BufRead::read_line. If we don’t do this
then it’s really awkward to use the stream when it is stored in a parent
struct because the reference returned by the line buffer makes the
mutable reference to self stay alive as well. That means the parent
struct can’t take any more references to itself while processing the
line.

The temporary C wrapper has been modified to return a different struct
that combines the stream with a buffer so that it can continue to return
a pointer to the owned string. This laughs in the face of the borrow
checker but it’s only temporary until the Rust conversion is complete
and it would be really difficult to pass a growable buffer from C to
Rust.