vbo: Make the parser struct public
That way the vbo can be constructed line-by-line without having to
concatenate all of the source into a single string. The line number
field of the Error struct has been removed because it will be easier to
use the line number from the script parser in the final error mesasges
instead.