vulkan_funcs: Add a unit test helper to override vkGetInstanceProcAddr
For the existing helper function that makes a fake Vulkan driver,
instead of directly constructing an empty Library struct, there is now a
more general helper function that overrides the function used for
vkGetInstanceProcAddr in the next call to Library::new. The original
helper function now uses this. The new helper function is also public so
unit tests can use it to test code that creates its own Library.